+++ CHANGES.md
@@ -10,6 +10,8 @@ The released versions correspond to PyPi releases.
   (see [#614](../../issues/614))
 * do not import pandas and related modules if it is not patched
   (see [#627](../../issues/627))
+* handle `pathlib.Path.owner()` and `pathlib.Path.group` by returning 
+  the current user/group name (see [#629](../../issues/629))
 
 ### Infrastructure
 * added test dependency check (see [#608](../../issues/608))